+++ aksel.nav.no/website/components/website-modules/seo/SEO.tsx
@@ -59,46 +59,3 @@ export function SEO({
     </Head>
   );
 }
-
-export function BaseSEO({ path }: { path: string }) {
-  return (
-    <Head>
-      <meta name="viewport" content="initial-scale=1.0, width=device-width" />
-      <meta
-        property="og:url"
-        content={`https://aksel.nav.no${path.split("?")[0].split("#")[0]}`}
-        key="ogurl"
-      />
-      <link rel="icon" href="/favicon.svg" sizes="any" type="image/svg+xml" />
-      <link
-        rel="shortcut icon"
-        href="/favicon.svg"
-        sizes="any"
-        type="image/svg+xml"
-      />
-      <meta property="og:site_name" content="Aksel" key="ogsitename" />
-    </Head>
-  );
-}
-
-/*
-
-<Head>
-
-        <meta
-          property="og:image"
-          content={
-            page?.seo?.image
-              ? urlFor(page?.seo?.image)
-                  .width(1200)
-                  .height(630)
-                  .fit("crop")
-                  .quality(100)
-                  .url()
-              : ""
-          }
-          key="ogimage"
-        />
-      </Head>
-
-*/
