+++ aksel.nav.no/website/pages/god-praksis/artikler/index.tsx
@@ -11,7 +11,7 @@ import { Button, ErrorMessage, Heading } from "@navikt/ds-react";
 import Footer from "components/layout/footer/Footer";
 import { Header } from "components/layout/header/Header";
 import ArtikkelCard from "components/sanity-modules/cards/ArtikkelCard";
-import { SEO } from "components/website-modules/SEO";
+import { SEO } from "components/website-modules/seo/SEO";
 import { AkselCubeStatic } from "components/website-modules/cube";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
