+++ aksel.nav.no/website/pages/god-praksis/index.tsx
@@ -16,7 +16,7 @@ import { Heading } from "@navikt/ds-react";
 import { Header } from "components/layout/header/Header";
 import ArtikkelCard from "components/sanity-modules/cards/ArtikkelCard";
 import GodPraksisCard from "components/sanity-modules/cards/GodPraksisCard";
-import { SEO } from "components/website-modules/SEO";
+import { SEO } from "components/website-modules/seo/SEO";
 import { AkselCubeStatic } from "components/website-modules/cube";
 import { Suspense, lazy } from "react";
