+++ aksel.nav.no/website/pages/komponenter/[...slug].tsx
@@ -24,7 +24,7 @@ import { Header } from "components/layout/header/Header";
 import { WithSidebar } from "components/layout/WithSidebar";
 import ComponentOverview from "components/sanity-modules/ComponentOverview";
 import IntroSeksjon from "components/sanity-modules/IntroSeksjon";
-import { SEO } from "components/website-modules/SEO";
+import { SEO } from "components/website-modules/seo/SEO";
 import { StatusTag } from "components/website-modules/StatusTag";
 import { SuggestionBlock } from "components/website-modules/suggestionblock";
 import { lazy, Suspense } from "react";
