+++ aksel.nav.no/website/pages/komponenter/index.tsx
@@ -23,7 +23,7 @@ import Footer from "components/layout/footer/Footer";
 import { Header } from "components/layout/header/Header";
 import ComponentOverview from "components/sanity-modules/ComponentOverview";
 import { IntroCards } from "components/website-modules/IntroCards";
-import { SEO } from "components/website-modules/SEO";
+import { SEO } from "components/website-modules/seo/SEO";
 import { Suspense, lazy } from "react";
 import { komponentKategorier } from "../../sanity/config";
